requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
httpx==0.27.0
gspread==6.1.2
google-auth==2.29.0
anthropic==0.28.0
openai==1.30.0          # optionnel — transcription vocale Whisper
python-multipart==0.0.9
